+++ src/containers/Orders/Item.tsx
@@ -22,20 +22,24 @@ const OrderItemComponent: React.FC<OrderItemComponentProps> = ({ cartItem }) => 
   return (
     <Grid borderBottom={`1px solid ${theme.colors.background}`} height="90px">
       <Grid gridTemplateColumns="1fr 1fr">
-        <Flex>
-          <Text fontWeight={600} fontSize={12}>
-            {name}
-          </Text>
-          <Tag
-            alignSelf="start"
-            justifySelf="start"
-            fontSize={11}
-            ml={2}
-            size="sm"
-            background="#c9cfd4"
-          >
-            {quantity}
-          </Tag>
+        <Flex flexDirection="column">
+          <Flex>
+            <Text fontWeight={600} fontSize={12}>
+              {name}
+            </Text>
+            <Tag
+              alignSelf="start"
+              justifySelf="start"
+              fontSize={11}
+              ml={2}
+              size="sm"
+              background="#c9cfd4"
+            >
+              {quantity}
+            </Tag>
+          </Flex>
+          <Text fontSize={12}>{`${quantity} units`}</Text>
+          <Text mt={1} fontSize={12}>{`Delivery fee not included`}</Text>
         </Flex>
         <Flex justifySelf="end">
           <Text fontWeight={600} fontSize={14}>{`${currency} ${total}.00`}</Text>