+++ src/containers/Returns/OrderReturnWeb.tsx
@@ -3,9 +3,12 @@ import * as React from 'react'
 import dayjs from 'dayjs'
 import RelativeTime from 'dayjs/plugin/relativeTime'
 
+import { get } from 'lodash'
 import { Form, Formik } from 'formik'
+import { ArrowLeft } from 'react-feather'
+import { ApolloError } from 'apollo-boost'
 import { useHistory } from 'react-router-dom'
-import { Flex, Grid, Button, Spinner } from '@chakra-ui/core'
+import { Flex, Grid, Button, Spinner, Tag, Image, useToast } from '@chakra-ui/core'
 
 import NoData from '../Checkout/NoDataScreen'
 
@@ -13,13 +16,117 @@ import { theme } from '../../theme'
 import { OrderReturnsProps } from '.'
 import { PageWrap } from '../../layouts'
 import { H3, Text } from '../../typography'
-import { Order } from '../../generated/graphql'
-import { ConnectedFormGroup, ConnectedTextArea } from '../../components/FormElements'
+import { Order, Maybe, Product, useRequestForReturnMutation } from '../../generated/graphql'
+import {
+  ConnectedFormGroup,
+  ConnectedTextArea,
+  ConnectedSelect
+} from '../../components/FormElements'
+import { formatError } from '../../utils'
+import { ERROR_TOAST, SUCCESS_TOAST } from '../../constants'
 
 dayjs.extend(RelativeTime)
 
+type Values = {
+  returnOrderId: Maybe<string> | undefined
+  returnProductName?: string
+  returnReason?: string
+  returnAction?: string
+}
+
+type OrderReturnProductProps = {
+  refundable: boolean
+  orderDate: string
+  product: Maybe<Product> | undefined
+  productToReturn?: string
+  setProductToReturn?: React.Dispatch<React.SetStateAction<Maybe<Product> | undefined>>
+}
+
+const OrderReturnProduct: React.FC<OrderReturnProductProps> = ({
+  product,
+  orderDate,
+  refundable,
+  productToReturn,
+  setProductToReturn
+}) => {
+  const maxSellCost = get(product, 'maxSellCost') as number
+  const tradeFedCost = get(product, 'tradeFedCost') as number
+
+  const discount = Math.round(((maxSellCost - tradeFedCost) / maxSellCost) * 100)
+
+  const handleProductClicked = () => {
+    if (setProductToReturn) {
+      setProductToReturn(product)
+    }
+  }
+  return (
+    <Flex
+      mt={3}
+      width={'100%'}
+      alignItems="center"
+      position="relative"
+      onClick={handleProductClicked}
+      justifyContent="space-between"
+      cursor={refundable ? 'pointer' : 'not-allowed'}
+      background={`${productToReturn === product?.id ? theme.colors.background : ''}`}
+    >
+      <Flex
+        width="100%"
+        borderRadius={3}
+        height="100px"
+        flexDirection="row"
+        border={`1px solid ${theme.colors.background}`}
+      >
+        <Flex width="50%" position="relative">
+          <Image
+            width="75%"
+            height="100%"
+            borderTopLeftRadius={3}
+            borderBottomLeftRadius={3}
+            src={product?.coverImage?.url || ''}
+            style={{ filter: refundable ? '' : 'grayscale(100%)' }}
+          />
+          {discount ? (
+            <Flex
+              top={0}
+              left={117}
+              width="50px"
+              height="50px"
+              position="absolute"
+              alignItems="center"
+              flexDirection="column"
+              justifyContent="center"
+              bg={refundable ? 'accent.700' : '#acacac'}
+            >
+              <Text color="white" fontSize="14px">
+                Save
+              </Text>
+              <Text color="white" fontSize="14px" fontWeight={600}>
+                {`${discount}%`}
+              </Text>
+            </Flex>
+          ) : null}
+        </Flex>
+        <Flex width={`100%`} ml={'-2rem'} flexDirection="column">
+          <Text my={2} color={`${refundable ? '' : '#acacac'}`} fontSize="14px" fontWeight={600}>
+            {product?.name}
+          </Text>
+          <Text my={2} color={`${refundable ? '' : '#acacac'}`} fontSize="12px">
+            {`Ordered: ${dayjs(orderDate).format('DD/MM/YYYY')} (${dayjs(orderDate).fromNow()})`}
+          </Text>
+          <Text color={`${refundable ? '' : '#acacac'}`} fontSize="12px">
+            {`${product?.currency} ${product?.tradeFedCost}.00`}
+          </Text>
+        </Flex>
+      </Flex>
+    </Flex>
+  )
+}
+
 const OrderReturns: React.FC<OrderReturnsProps> = ({
   orders,
+  actions,
+  reasons,
   fetchingOrders,
   pastOrders,
   activeOrders,
@@ -29,7 +136,16 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
   noActiveOrdersCaption
 }) => {
   const history = useHistory()
-  const [currentPage, setCurrentPage] = React.useState<string>('active')
+  const toast = useToast()
+
+  const [values, setValues] = React.useState<Values>()
+  const [orderToReturn, setOrderToReturn] = React.useState<Order>()
+  const [productToReturn, setProductToReturn] = React.useState<Maybe<Product> | undefined>()
+  const [currentPage, setCurrentPage] = React.useState<string>('past')
+  const [currentAction, setCurrentAction] = React.useState<string>('')
+  const [currentReason, setCurrentReason] = React.useState<string>('')
+  const [selectedOrderToReturn, setSelectedOrderToReturn] = React.useState<boolean>()
+
   const cancelStyles = {
     cursor: 'pointer',
     textDecoration: 'underline',
@@ -41,6 +157,54 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
   const handleOrderHistoryClicked = () => {
     history.push('/orders')
   }
+  const handleOrderClicked = (id: string | null | undefined) => {
+    const theOrder = pastOrders?.filter((order) => order?.orderNumber === id)[0]
+    const formValues = {
+      ...values,
+      returnOrderId: theOrder?.orderNumber
+    }
+    setValues(formValues)
+    setOrderToReturn(theOrder)
+    setSelectedOrderToReturn(true)
+  }
+
+  const goBackToOrders = () => {
+    setSelectedOrderToReturn(false)
+  }
+
+  const refundableProducts = orderToReturn?.items?.filter((order) => order?.product?.isRefundable)
+  const nonRefundableProducts = orderToReturn?.items?.filter(
+    (order) => !order?.product?.isRefundable
+  )
+
+  const isNonRefundableEmpty = nonRefundableProducts?.length
+
+  const message = `Hi TradeFed, I would like to request a return of ${productToReturn?.name} from Order# ${orderToReturn?.orderNumber}. I have indicated my reason and preferred action. Looking forward to hearing from you. Thanks in advance!`
+  const returnComment = (orderToReturn && productToReturn && message) ?? ''
+
+  const handleReasonChanged = (event: React.ChangeEvent<HTMLSelectElement>) => {
+    event.persist()
+    const rsn = event?.target?.value
+    setCurrentReason(rsn)
+  }
+
+  const handleActionChanged = (event: React.ChangeEvent<HTMLSelectElement>) => {
+    event.persist()
+    const actn = event?.target?.value
+    setCurrentAction(actn)
+  }
+
+  const [requestForReturn, { loading: returnLoading }] = useRequestForReturnMutation({
+    onError: (err: ApolloError) => toast({ description: err.message, ...ERROR_TOAST }),
+    onCompleted: ({ placeItemReturnRequest }) => {
+      const msg = placeItemReturnRequest?.payload as string
+      toast({ description: msg, ...SUCCESS_TOAST })
+      setTimeout(() => {
+        history.push('/orders')
+      }, 1000)
+    }
+  })
+
   return (
     <PageWrap title="Order Returns" alignSelf="center">
       <Grid
@@ -57,7 +221,7 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
             <H3 textAlign="left" fontSize={18} fontWeight={600}>
               Return an Order
             </H3>
-            <Text onClick={handleOrderHistoryClicked} fontSize={12} style={cancelStyles}>
+            <Text onClick={handleOrderHistoryClicked} fontSize={14} style={cancelStyles}>
               Order History
             </Text>
           </Flex>
@@ -70,7 +234,6 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
           borderRadius={5}
           background={theme.colors.accent[50]}
           boxShadow={theme.boxShadowMedium}
-          height="400px"
         >
           <Flex flexDirection="column" width="100%">
             <Flex justify="space-between" mb={5}>
@@ -102,46 +265,78 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
               </Button>
             </Flex>
             <Formik
-              initialValues={{ signatoryName: '', signatoryRelation: '' }}
-              onSubmit={() => {}}
+              initialValues={{}}
+              onSubmit={async (items, { setStatus, setSubmitting }) => {
+                const input = {
+                  input: {
+                    action: String(currentAction),
+                    reason: String(currentReason),
+                    product: productToReturn?.id as string,
+                    comment: returnComment,
+                    orderNumber: orderToReturn?.orderNumber as string
+                  }
+                }
+                setStatus(null)
+                try {
+                  setSubmitting(true)
+                  requestForReturn({
+                    variables: input
+                  })
+                  setSubmitting(false)
+                } catch (error) {
+                  setStatus(formatError(error))
+                }
+              }}
             >
-              <Form style={{ width: '100%' }}>
-                <ConnectedFormGroup
-                  label="Order ID*"
-                  name="orderId"
-                  type="text"
-                  placeholder="Eg. TFSA-aw3erdfsd"
-                />
-                <ConnectedFormGroup
-                  label="Delivery Address Name"
-                  name="addressName"
-                  type="text"
-                  placeholder="Eg. Mum's Place"
-                />
-                <Flex flexDirection="column">
-                  <select placeholder="Why would like to return the order?">
-                    <option>Why would like to return the order?</option>
-                    <option>1 don't care</option>
-                    <option>It was a mistaken purchase</option>
-                    <option>1</option>
-                    <option>1</option>
-                  </select>
-                  <select
-                    style={{ marginTop: 10, marginBottom: 10 }}
-                    placeholder="What would you like to be done?"
+              {({ isSubmitting, status }) => (
+                <Form style={{ width: '100%' }}>
+                  <ConnectedFormGroup
+                    type="text"
+                    label="Order ID*"
+                    name="returnOrderId"
+                    placeholder="Eg. TFSA-aw3erdfsd"
+                    value={`${orderToReturn?.orderNumber || ''}`}
+                  />
+                  <ConnectedFormGroup
+                    type="text"
+                    label="Product Name"
+                    name="returnProductName"
+                    placeholder="Eg. Ngala Lamp Stand"
+                    value={`${productToReturn ? `${productToReturn?.name}` : ''}`}
+                  />
+                  <ConnectedSelect
+                    options={reasons}
+                    onChange={handleReasonChanged}
+                    name="returnReason"
+                    label="Why would like to return the order?* "
+                    placeholder="Select reason for return"
+                  />
+                  <ConnectedSelect
+                    options={actions}
+                    onChange={handleActionChanged}
+                    name="returnAction"
+                    label="What would you like to be done?* "
+                    placeholder="Select preferred action to be taken"
+                  />
+                  <ConnectedTextArea
+                    minHeight={`180px`}
+                    name="returnComment"
+                    value={returnComment}
+                    handleSetTags={() => {}}
+                    label="Comment(Optional)"
+                    placeholder="Comments (Any additional information)"
+                  />
+                  <Button
+                    mt={4}
+                    width="100%"
+                    type="submit"
+                    variantColor="brand"
+                    isLoading={returnLoading}
                   >
-                    <option>What would you like to be done?</option>
-                    <option>Exchange the order</option>
-                    <option>Something else</option>
-                    <option>1</option>
-                    <option>1</option>
-                  </select>
-                </Flex>
-                <ConnectedTextArea name="returnReason" handleSetTags={() => {}} />
-                <Button mt={4} width="100%" type="submit" variantColor="brand">
-                  RETURN
-                </Button>
-              </Form>
+                    RETURN
+                  </Button>
+                </Form>
+              )}
             </Formik>
           </Flex>
         </Flex>
@@ -156,12 +351,12 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
           flexDirection="column"
         >
           {activeOrderPage && activeOrders?.length ? (
-            <React.Fragment>
+            <Flex flexDirection="column">
               <Text fontWeight={600} fontSize={14}>
                 {`All Active Orders (${activeOrders ? activeOrders?.length : ''}`})
               </Text>
-              <Flex flexDirection="column">
-                <Grid mt={3} p={4}>
+              <Flex flexDirection="column" width={`100%`}>
+                <Grid mt={3}>
                   <Flex
                     justify="space-between"
                     mb={3}
@@ -188,7 +383,7 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
                   ))}
                 </Grid>
               </Flex>
-            </React.Fragment>
+            </Flex>
           ) : (
             activeOrderPage &&
             !activeOrders?.length && (
@@ -196,39 +391,112 @@ const OrderReturns: React.FC<OrderReturnsProps> = ({
             )
           )}
           {pastOrderPage && pastOrders?.length ? (
-            <React.Fragment>
-              <Text fontWeight={600} fontSize={14}>
-                {`All Past Orders (${pastOrders ? pastOrders?.length : ''}`})
-              </Text>
-              <Flex flexDirection="column">
-                <Grid mt={3} p={4}>
-                  <Flex
-                    justify="space-between"
-                    mb={3}
-                    borderBottom={`1px dashed ${theme.colors.background}`}
-                    pb={3}
-                  >
-                    <Text fontWeight={600}>Order#</Text>
-                    <Text fontWeight={600}>Delivery Date</Text>
-                    <Text fontWeight={600}>Delivered To</Text>
+            <Flex flexDirection="column">
+              {selectedOrderToReturn ? (
+                <React.Fragment>
+                  <Flex cursor="pointer" onClick={goBackToOrders}>
+                    <ArrowLeft />
+                    <Text fontWeight={600} fontSize={14} ml={3}>
+                      {`Back to Orders`}
+                    </Text>
                   </Flex>
-                  {fetchingOrders && <Spinner margin="auto" />}
-                  {pastOrders?.map((order: Order, index: number) => (
+                  <Flex p={3} mt={2} flexDirection="column">
+                    <Text fontWeight={600} fontSize={14}>
+                      {`Order# ${orderToReturn?.orderNumber}`}
+                    </Text>
                     <Flex
-                      borderBottom={`1px dashed ${theme.colors.background}`}
-                      textAlign="left"
-                      key={`${index}_order_item`}
-                      justify="space-between"
-                      py={4}
+                      pb={5}
+                      borderBottom={`${isNonRefundableEmpty ? '1px dashed #acacac' : ''}`}
+                      flexDirection="column"
+                      maxHeight="250px"
+                      overflowY="scroll"
                     >
-                      <Text fontSize={12}>{order?.orderNumber}</Text>
-                      <Text fontSize={12}>{dayjs(order?.deliveryDate).fromNow()}</Text>
-                      <Text fontSize={12}>{order?.deliveryAddress?.name}</Text>
+                      {refundableProducts?.map((item, index) => {
+                        return (
+                          <OrderReturnProduct
+                            refundable
+                            product={item?.product}
+                            key={`${index}_order_product`}
+                            orderDate={orderToReturn?.created_at}
+                            productToReturn={productToReturn?.id}
+                            setProductToReturn={setProductToReturn}
+                          />
+                        )
+                      })}
                     </Flex>
-                  ))}
-                </Grid>
-              </Flex>
-            </React.Fragment>
+                    {isNonRefundableEmpty ? (
+                      <Flex flexDirection="column" mt={2}>
+                        <Flex>
+                          <Tag
+                            fontSize={12}
+                            mt={3}
+                            size="sm"
+                            justifySelf="start"
+                            background={theme.colors.tag}
+                            color={theme.colors.tagText}
+                          >
+                            NON-REFUNDABLE ITEMS
+                          </Tag>
+                        </Flex>
+                        <Flex mt={3}>
+                          {nonRefundableProducts?.map((item, index) => {
+                            return (
+                              <OrderReturnProduct
+                                key={`${index}_order_product_non`}
+                                product={item?.product}
+                                refundable={false}
+                                orderDate={orderToReturn?.created_at}
+                              />
+                            )
+                          })}
+                        </Flex>
+                      </Flex>
+                    ) : null}
+                  </Flex>
+                </React.Fragment>
+              ) : (
+                <React.Fragment>
+                  <Text fontWeight={600} fontSize={14}>
+                    {`All Past Orders (${pastOrders ? pastOrders?.length : ''}`})
+                  </Text>
+                  <Flex flexDirection="column">
+                    <Grid mt={3} p={4}>
+                      <Flex
+                        justify="space-between"
+                        mb={3}
+                        borderBottom={`1px dashed ${theme.colors.background}`}
+                        pb={3}
+                      >
+                        <Text fontWeight={600}>Order#</Text>
+                        <Text fontWeight={600}>Delivery Date</Text>
+                        <Text fontWeight={600}>Delivered To</Text>
+                      </Flex>
+                      {fetchingOrders && <Spinner margin="auto" />}
+                      {pastOrders?.map((order: Order, index: number) => (
+                        <Flex
+                          borderBottom={`1px dashed ${theme.colors.background}`}
+                          textAlign="left"
+                          key={`${index}_order_item`}
+                          justify="space-between"
+                          py={4}
+                        >
+                          <Flex style={{ textDecoration: 'underline' }} cursor="pointer">
+                            <Text
+                              fontSize={12}
+                              onClick={() => handleOrderClicked(order?.orderNumber)}
+                            >
+                              {order?.orderNumber}
+                            </Text>
+                          </Flex>
+                          <Text fontSize={12}>{dayjs(order?.deliveryDate).fromNow()}</Text>
+                          <Text fontSize={12}>{order?.deliveryAddress?.name}</Text>
+                        </Flex>
+                      ))}
+                    </Grid>
+                  </Flex>
+                </React.Fragment>
+              )}
+            </Flex>
           ) : (
             pastOrderPage &&
             !pastOrders?.length && (